Create Snake Eating.cpp
The Chef doesn't care about snakes eating each other. All he wants is to have as many snakes as possible, which are at least some particular lengths long. You are given Q values: K1, K2, .., KQ. Treat each of them independently and output the answer for each. That is, for each Ki, assume that you start out from the beginning with all the snakes alive and the lengths as the initial values given in the input, and find out the maximum number of snakes you can get whose length is at least Ki.